+++ src/main/java/com/kalsym/product/service/GenerateSiteMapScheduler.java
@@ -1,30 +1,19 @@
 package com.kalsym.product.service;
 
-import com.kalsym.product.service.ProductServiceApplication;
-import com.kalsym.product.service.repository.LocationConfigRepository;
-import com.kalsym.product.service.repository.StoreRepository;
-import com.kalsym.product.service.utility.Logger;
-import com.kalsym.product.service.model.LocationConfig;
-import com.kalsym.product.service.model.store.Store;
+import com.kalsym.product.service.service.SiteMapService;
 
 import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 @Component
 public class GenerateSiteMapScheduler {
     
     @Autowired
-    LocationConfigRepository locationConfigRepository;
-    
-    @Value("${marketplace.url}")
-    private String marketPlaceUrl;
+    SiteMapService siteMapService;
 
     @Value("${path.main.sitemapxml}")
     private String pathMainXml;
@@ -36,46 +25,15 @@ public class GenerateSiteMapScheduler {
     @Scheduled(cron = "0 0 23 * * *")
     public void generateSitemapXml() throws Exception{
 
-        List<LocationConfig> locationListing = locationConfigRepository.findAll();
+        File f = new File(pathMainXml);
 
-        String mainXml = 
-        "<?xml version=\"1.0\" encoding=\"UTF-8\"?>"
-        +"\n<urlset xmlns=\"http://www.sitemaps.org/schemas/sitemap/0.9\">"
-        ;
-        String  locationXml = "";
-        String endXml = "\n</urlset>";
+        if(f.exists() && !f.isDirectory()) { 
+            String finalXml = siteMapService.generateLocationSitemap();
 
-        for(LocationConfig lc:locationListing){
-
-            locationXml += 
-            "\n\t<url>"
-            +"\n\t\t<loc>"+marketPlaceUrl+"/location/"+lc.getCityId()+"</loc>"
-            +"\n\t\t<lastmod>2022-07-26T08:56:02+00:00</lastmod>"
-            +"\n\t</url>"
-            ;
-
+            siteMapService.overWriteFile(finalXml);
         }
 
-        String finalXml = mainXml+locationXml+endXml;
-
-        // System.out.println("CHECKING THE GENERATE SITEMAP ::::::"+finalXml);
-        
-        //delete first
-        File fold = new File(pathMainXml);
-        fold.delete();
-
-        //then wrtite
-        File fnew = new File(pathMainXml);
-
-        try {
-            FileWriter f2 = new FileWriter(fnew, false);
-            f2.write(finalXml);
-            f2.close();
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }   
-
+         
     }