+++ src/main/java/com/kalsym/product/service/controller/SitemapController.java
@@ -9,6 +9,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.tomcat.util.http.fileupload.IOUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -21,27 +23,54 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.kalsym.product.service.service.SiteMapService;
+
 @Controller
 @RestController()
 
 public class SitemapController {
 
+    @Value("${path.main.sitemapxml}")
+    private String pathMainXml;
+
+    @Autowired
+    SiteMapService siteMapService;
+
     @GetMapping(path = {"/sitemap.xml"}, produces = {MediaType.APPLICATION_XML_VALUE,MediaType.TEXT_XML_VALUE}, consumes = MediaType.ALL_VALUE)
     public ResponseEntity<?> serverFileList(HttpServletRequest request) throws IOException {
 
-        byte[] fileContent = Files.readAllBytes(Paths.get("C:/Users/Nur Iman/Desktop/Projects/java/product-service/src/main/resources/static/sitemap.xml"));
+        File f = new File(pathMainXml);
+
+        if(f.exists() && !f.isDirectory()) { 
+
+            byte[] fileContent = Files.readAllBytes(Paths.get(pathMainXml));
         
-        HttpHeaders responseHeaders = new HttpHeaders();
-        responseHeaders.setContentType(MediaType.APPLICATION_XML);
-        responseHeaders.setContentLength(fileContent.length);
+            HttpHeaders responseHeaders = new HttpHeaders();
+            responseHeaders.setContentType(MediaType.APPLICATION_XML);
+            responseHeaders.setContentLength(fileContent.length);
+    
+            return new ResponseEntity<byte[]>(fileContent, responseHeaders,
+                    HttpStatus.OK);
 
-        return new ResponseEntity<byte[]>(fileContent, responseHeaders,
-                HttpStatus.OK);
-        // ResponseEntity<byte[]> response = new ResponseEntity<>(fileContent)
-        // response.setStatus(HttpStatus.NOT_MODIFIED.value());
-        // response.getOutputStream().write(fileContent);
-        // return response;
+        }
+        
+        else{
 
+            System.out.println("OPPPSSSS FILE NOT FOUND ::::::::");
+
+            siteMapService.createFileSitemapLocation();
+            String finalXml = siteMapService.generateLocationSitemap();
+            siteMapService.overWriteFile(finalXml);
+
+            byte[] byteArr = finalXml.getBytes();
+
+            HttpHeaders responseHeaders = new HttpHeaders();
+            responseHeaders.setContentType(MediaType.APPLICATION_XML);
+    
+            return new ResponseEntity<byte[]>(byteArr,responseHeaders,
+                    HttpStatus.OK);
+
+        }
       
     }